[feat] 게시판 복수 조회 시, 계산 비용 문제에 대한 문제 구현 - 게시판 복수 조회 시, 각 게시판의 게시글, 댓글개수 계산해서 가져오도록 구현

+++ src/main/java/com/totalhubsite/backend/domain/board/dto/response/BoardDetailResponseDto.java
@@ -9,15 +9,15 @@ public record BoardDetailResponseDto(
     Long id,
     String name,
     String description,
-    Integer postCount,
-    Integer commentCount,
+    Long postCount,
+    Long commentCount,
     MemberInfoResponseDto member
 
 ) {
 
     @Builder
-    public BoardDetailResponseDto(Long id, String name, String description, Integer postCount,
-        Integer commentCount, MemberInfoResponseDto member) {
+    public BoardDetailResponseDto(Long id, String name, String description, Long postCount,
+        Long commentCount, MemberInfoResponseDto member) {
         this.id = id;
         this.name = name;
         this.description = description;
@@ -40,7 +40,7 @@ public record BoardDetailResponseDto(
             .build();
     }
 
-    public static BoardDetailResponseDto fromEntity(Board entity, Integer postCount, Integer commentCount) {
+    public static BoardDetailResponseDto fromEntity(Board entity, Long postCount, Long commentCount) {
         Member findMember = entity.getMember();
         MemberInfoResponseDto memberInfo = MemberInfoResponseDto.fromEntity(findMember);
